+++ lib/Oystercard.rb
@@ -1,6 +1,6 @@
 class Oystercard
 
-    attr_reader :balance, :entry_station, :journeys, :journey
+    attr_reader :balance, :entry_station, :journey
     attr_accessor :limit
 
     DEFAULT = 90
@@ -9,7 +9,7 @@ class Oystercard
         @balance = 0
         @limit = limit
         @min = min
-        @journeys = []
+
     end
 
     def top_up(num)
@@ -17,22 +17,22 @@ class Oystercard
       @balance > limit ? raise("maximum balance reached #{limit}!") : @balance
     end
 
-    def in_journey?
-      @entry_station != nil
-    end
-
+    # def in_journey?
+    #   @entry_station != nil
+    # end
+    #
     def touch_in(num, station)
         if num < @min then raise("not enough funds")
         end
         @entry_station = station
     end
-
-    def touch_out(num, station)
-        deduct(num)
-        @exit_station = station
-        @journey = { @entry_station => @exit_station }
-        @entry_station = nil      
-    end
+    #
+    # def touch_out(num, station)
+    #     deduct(num)
+    #     @exit_station = station
+    #     @journey = { @entry_station => @exit_station }
+    #     @entry_station = nil
+    # end
 
     private
     def deduct(num)
